Changed some limits reguarding the max length of messages.
Hastebin upload limit for "hastebin" and "compile" commands: 10,000 -> 20,000, some outputs are a bit longer and them being cut off is a bit annoying for some.

Message length for "compile" when responding to user: 1988 -> 1250, I found it was easy to abuse the compile command to flood chat making this change just for preventing really long messages.

Messange length for "hastebin" when responding to user: 500 -> 450, reduced slightly as the code preview doesn't need to be that long.